% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_stats.r
\name{anova1}
\alias{anova1}
\title{One-way ANOVA}
\usage{
anova1(Y, roi = NULL)
}
\arguments{
\item{Y}{list of data matrices, one per group}

\item{roi}{region of interest mask}
}
\value{
SPM_F or SPM0D_F object
}
\description{
One-way ANOVA
}
\examples{
Y1 <- matrix(rnorm(30), 3, 10)
Y2 <- matrix(rnorm(30, mean = 0.5), 3, 10)
Y3 <- matrix(rnorm(30, mean = 1), 3, 10)
spm <- anova1(list(Y1, Y2, Y3))
spmi <- inference(spm, alpha = 0.05, withBonf = TRUE)
}
