% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_stats.r
\name{anova1rm}
\alias{anova1rm}
\title{One-way repeated measures ANOVA}
\usage{
anova1rm(Y, roi = NULL)
}
\arguments{
\item{Y}{list of data matrices, one per condition (same subjects)}

\item{roi}{region of interest mask}
}
\value{
SPM_F or SPM0D_F object
}
\description{
One-way repeated measures ANOVA
}
