% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_stats.r
\name{anova2}
\alias{anova2}
\title{Two-way ANOVA (factorial design)}
\usage{
anova2(Y, A, B, roi = NULL)
}
\arguments{
\item{Y}{data matrix (J x Q)}

\item{A}{factor A (length J vector)}

\item{B}{factor B (length J vector)}

\item{roi}{region of interest mask}
}
\value{
list of SPM_F objects for main effects and interaction
}
\description{
Two-way ANOVA (factorial design)
}
