% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_rft.r
\name{cluster_pvalues}
\alias{cluster_pvalues}
\title{Calculate cluster p-values}
\usage{
cluster_pvalues(clusters, fwhm, resels)
}
\arguments{
\item{clusters}{cluster data frame}

\item{fwhm}{field smoothness}

\item{resels}{resel counts}
}
\description{
Calculate cluster p-values
}
