% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_classes.R
\name{dflist2str}
\alias{dflist2str}
\title{Convert degrees of freedom to string}
\usage{
dflist2str(df)
}
\arguments{
\item{df}{degrees of freedom (vector)}
}
\description{
Convert degrees of freedom to string
}
