% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_rft.r, R/spm1d_r_stats.r,
%   R/spm_classes.R
\docType{data}
\name{eps}
\alias{eps}
\title{spm1d R Package - Random Field Theory (RFT) Module}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
eps

eps

eps
}
\description{
Implements Random Field Theory calculations for 1D statistical inference
Based on Worsley et al. (1996) and Friston et al. (1994)

Implements main statistical tests: GLM, t-tests, ANOVA, regression

One-Dimensional Statistical Parametric Mapping in R
Port of Python spm1d package by Todd Pataky
}
\details{
This module implements the core SPM object classes
Machine epsilon for numerical stability
}
\keyword{datasets}
