% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_geom.r
\name{estimate_fwhm}
\alias{estimate_fwhm}
\title{spm1d R Package - Geometry Module}
\usage{
estimate_fwhm(residuals)
}
\arguments{
\item{residuals}{residuals from model fit (J x Q matrix)}
}
\value{
estimated FWHM value
}
\description{
Estimates field smoothness (FWHM) and resel counts for Random Field Theory
Estimate Full-Width at Half-Maximum (FWHM) from residuals
}
\details{
FWHM is estimated from the residuals using the variance of partial
derivatives (Worsley et al., 1992). This provides a measure of the
smoothness of the statistical field.

FWHM = sqrt(4 * log(2) / lambda)
where lambda is the roughness parameter estimated from residual derivatives
}
