% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_geom.r
\name{estimate_fwhm_alt}
\alias{estimate_fwhm_alt}
\title{Estimate FWHM using alternative method (for validation)}
\usage{
estimate_fwhm_alt(residuals)
}
\arguments{
\item{residuals}{residuals from model fit}
}
\description{
Estimate FWHM using alternative method (for validation)
}
