% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_rft.r
\name{euler_characteristic}
\alias{euler_characteristic}
\title{Euler characteristic for different distributions}
\usage{
euler_characteristic(u, df, stat = "T")
}
\arguments{
\item{u}{threshold value}

\item{df}{degrees of freedom}

\item{stat}{statistic type}
}
\description{
Euler characteristic for different distributions
}
