% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_geom.r
\name{expected_ec}
\alias{expected_ec}
\title{Calculate expected Euler characteristic for a field}
\usage{
expected_ec(threshold, df, stat_type, resels)
}
\arguments{
\item{threshold}{threshold value}

\item{df}{degrees of freedom}

\item{stat_type}{statistic type ('T', 'F', etc.)}

\item{resels}{resel counts}
}
\description{
Calculate expected Euler characteristic for a field
}
