% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_rft.r
\name{f_isf}
\alias{f_isf}
\title{F distribution: inverse survival function (1D) with Bonferroni correction}
\usage{
f_isf(alpha, df, fwhm, resels, withBonf = TRUE)
}
\arguments{
\item{alpha}{significance level}

\item{df}{degrees of freedom (vector: c(df_num, df_den))}

\item{fwhm}{field smoothness}

\item{resels}{resel counts}

\item{withBonf}{use Bonferroni correction if less severe (default TRUE)}
}
\description{
F distribution: inverse survival function (1D) with Bonferroni correction
}
