% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_rft.r
\name{find_clusters}
\alias{find_clusters}
\title{Find clusters in a 1D continuum that exceed threshold}
\usage{
find_clusters(z, zstar)
}
\arguments{
\item{z}{statistic continuum}

\item{zstar}{threshold}
}
\value{
data frame with cluster information
}
\description{
Find clusters in a 1D continuum that exceed threshold
}
