% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_stats.r
\name{glm}
\alias{glm}
\title{General Linear Model for t contrasts}
\usage{
glm(Y, X, c, Q = NULL, roi = NULL)
}
\arguments{
\item{Y}{dependent variable (J x Q matrix: J observations, Q nodes)}

\item{X}{design matrix (J x B: J observations, B parameters)}

\item{c}{contrast vector (length B)}

\item{Q}{non-sphericity specifiers (not currently supported)}

\item{roi}{region of interest mask (logical vector length Q)}
}
\value{
SPM_T or SPM0D_T object
}
\description{
General Linear Model for t contrasts
}
\examples{
# Example: two-sample t-test via GLM
Y <- rbind(matrix(rnorm(50), 5, 10), matrix(rnorm(50, 1), 5, 10))
X <- cbind(1, c(rep(0, 5), rep(1, 5)))
c <- c(0, 1)  # test second parameter
spm <- glm(Y, X, c)
spmi <- inference(spm, alpha = 0.05, two_tailed = TRUE, withBonf = TRUE)
}
