% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_geom.r
\name{gradient}
\alias{gradient}
\title{Calculate gradient (first derivative) of 1D data}
\usage{
gradient(x)
}
\arguments{
\item{x}{1D data vector or matrix}
}
\value{
gradient (same dimensions as input)
}
\description{
Calculate gradient (first derivative) of 1D data
}
