% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_inference.r
\name{inference}
\alias{inference}
\title{spm1d R Package - Inference Module}
\usage{
inference(
  spm,
  alpha = 0.05,
  two_tailed = TRUE,
  withBonf = TRUE,
  cluster_threshold = NULL
)
}
\arguments{
\item{spm}{SPM object (0D or 1D)}

\item{alpha}{significance level (default 0.05)}

\item{two_tailed}{logical, perform two-tailed test (for T statistics)}

\item{withBonf}{Logical, use Bonferroni correction}

\item{cluster_threshold}{Logical}
}
\description{
Implements statistical inference for SPM objects
Adds inference() methods to SPM classes
Perform statistical inference on SPM object
}
