% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_inference.r
\name{plot.SPM}
\alias{plot.SPM}
\title{Plot SPM object (before inference)}
\usage{
\method{plot}{SPM}(
  x,
  xlab = "Time / Position",
  ylab = NULL,
  main = NULL,
  col = "black",
  lwd = 2,
  ...
)
}
\arguments{
\item{x}{SPM object}

\item{xlab}{String x axis label}

\item{ylab}{String y axis label}

\item{main}{String Main plot title}

\item{col}{String}

\item{lwd}{Integer Line width}

\item{...}{additional plotting parameters}
}
\description{
Plot SPM object (before inference)
}
