% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_inference.r
\name{plot.SPMi}
\alias{plot.SPMi}
\title{Plot SPM inference results}
\usage{
\method{plot}{SPMi}(
  x,
  xlab = "Time / Position",
  ylab = NULL,
  main = NULL,
  plot_threshold = TRUE,
  plot_clusters = TRUE,
  col = "black",
  col_threshold = "red",
  col_cluster = "lightblue",
  lwd = 2,
  ...
)
}
\arguments{
\item{x}{SPMi object (inference result)}

\item{xlab}{String x axis label}

\item{ylab}{String y axis label}

\item{main}{String Main plot title}

\item{plot_threshold}{Logical}

\item{plot_clusters}{Logical}

\item{col}{String}

\item{col_threshold}{String}

\item{col_cluster}{String}

\item{lwd}{Integer Line width}

\item{...}{additional plotting parameters}
}
\description{
Plot SPM inference results
}
