% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_stats.r
\name{regress}
\alias{regress}
\title{Linear regression (correlation analysis)}
\usage{
regress(Y, x, roi = NULL)
}
\arguments{
\item{Y}{dependent variable (J x Q matrix)}

\item{x}{independent variable (length J vector)}

\item{roi}{region of interest mask}
}
\value{
SPM_T or SPM0D_T object
}
\description{
Linear regression (correlation analysis)
}
\examples{
x <- 1:10
Y <- outer(x, rep(1, 5)) + matrix(rnorm(50), 10, 5)
spm <- regress(Y, x)
spmi <- inference(spm, alpha = 0.05)
}
