% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_geom.r
\name{resel_counts_from_residuals}
\alias{resel_counts_from_residuals}
\title{Calculate resel counts from residuals for 1D SPM}
\usage{
resel_counts_from_residuals(residuals, fwhm)
}
\arguments{
\item{residuals}{matrix of residuals (J x Q)}

\item{fwhm}{estimated FWHM}
}
\value{
resel count
}
\description{
Calculate resel counts from residuals for 1D SPM
}
\details{
For 1D data, the standard formula is:
resels = (Q-1) / FWHM
where Q is the number of nodes
}
