% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_geom.r
\name{smooth_gaussian}
\alias{smooth_gaussian}
\title{Smooth 1D data with Gaussian kernel}
\usage{
smooth_gaussian(x, fwhm)
}
\arguments{
\item{x}{1D data vector or matrix (rows are observations)}

\item{fwhm}{Full-Width at Half-Maximum of smoothing kernel}
}
\value{
smoothed data
}
\description{
Smooth 1D data with Gaussian kernel
}
