% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_geom.r
\name{smooth_moving_average}
\alias{smooth_moving_average}
\title{Apply moving average smoothing}
\usage{
smooth_moving_average(x, window_size)
}
\arguments{
\item{x}{1D data}

\item{window_size}{size of moving average window}
}
\description{
Apply moving average smoothing
}
