% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_classes.R
\name{SPM}
\alias{SPM}
\title{Base 1D SPM Class}
\usage{
SPM(
  STAT,
  z,
  df,
  fwhm,
  resels,
  X = NULL,
  beta = NULL,
  residuals = NULL,
  sigma2 = NULL,
  roi = NULL
)
}
\arguments{
\item{STAT}{statistic type ('T', 'F', 'X2', 'T2')}

\item{z}{test statistic continuum (vector)}

\item{df}{degrees of freedom}

\item{fwhm}{full-width at half-maximum (field smoothness)}

\item{resels}{resel counts}

\item{X}{design matrix (optional)}

\item{beta}{fitted parameters (optional)}

\item{residuals}{residuals (optional)}

\item{sigma2}{variance (optional)}

\item{roi}{region of interest mask (optional)}
}
\description{
Base 1D SPM Class
}
