% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_classes.R
\name{SPM0D}
\alias{SPM0D}
\title{Base 0D SPM Class}
\usage{
SPM0D(STAT, z, df, beta = NULL, residuals = NULL, sigma2 = NULL)
}
\arguments{
\item{STAT}{statistic type ('T', 'F', 'X2', 'T2')}

\item{z}{test statistic value}

\item{df}{degrees of freedom}

\item{beta}{regression coefficients (optional)}

\item{residuals}{residuals (optional)}

\item{sigma2}{variance (optional)}
}
\description{
Base 0D SPM Class
}
