% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_classes.R
\name{SPM0Di}
\alias{SPM0Di}
\title{SPM Inference Result (0D)}
\usage{
SPM0Di(spm, alpha, zstar, p)
}
\arguments{
\item{spm}{original SPM object}

\item{alpha}{significance level}

\item{zstar}{critical threshold}

\item{p}{p-value}
}
\description{
SPM Inference Result (0D)
}
