% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_classes.R
\name{SPMi}
\alias{SPMi}
\title{SPM Inference Result (1D)}
\usage{
SPMi(spm, alpha, zstar, clusters, p_set = NULL, p_cluster = NULL)
}
\arguments{
\item{spm}{original SPM object}

\item{alpha}{significance level}

\item{zstar}{critical threshold}

\item{clusters}{cluster information}

\item{p_set}{p-value for set-level inference}

\item{p_cluster}{p-values for cluster-level inference}
}
\description{
SPM Inference Result (1D)
}
