% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_stats.r
\name{standardize_data}
\alias{standardize_data}
\title{Standardize data (mean 0, sd 1)}
\usage{
standardize_data(Y)
}
\arguments{
\item{Y}{data matrix}
}
\value{
standardized data
}
\description{
Standardize data (mean 0, sd 1)
}
