% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_stats.r
\name{ttest}
\alias{ttest}
\title{One-sample t-test}
\usage{
ttest(Y, mu = 0, roi = NULL)
}
\arguments{
\item{Y}{data matrix (J x Q: J observations, Q nodes)}

\item{mu}{null hypothesis mean (default 0)}

\item{roi}{region of interest mask}
}
\value{
SPM_T or SPM0D_T object
}
\description{
One-sample t-test
}
\examples{
Y <- matrix(rnorm(50, mean = 0.5), 5, 10)
spm <- ttest(Y)
spmi <- inference(spm, alpha = 0.05, two_tailed = TRUE)
}
