% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_stats.r
\name{ttest2}
\alias{ttest2}
\title{Two-sample t-test}
\usage{
ttest2(YA, YB, equal_var = TRUE, roi = NULL)
}
\arguments{
\item{YA}{data for group A (JA x Q matrix)}

\item{YB}{data for group B (JB x Q matrix)}

\item{equal_var}{assume equal variance (default TRUE)}

\item{roi}{region of interest mask}
}
\value{
SPM_T or SPM0D_T object
}
\description{
Two-sample t-test
}
\examples{
YA <- matrix(rnorm(50), 5, 10)
YB <- matrix(rnorm(50, mean = 1), 5, 10)
spm <- ttest2(YA, YB)
spmi <- inference(spm, alpha = 0.05)
}
