% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_stats.r
\name{ttest_paired}
\alias{ttest_paired}
\title{Paired t-test}
\usage{
ttest_paired(YA, YB, roi = NULL)
}
\arguments{
\item{YA}{data for condition A (J x Q matrix)}

\item{YB}{data for condition B (J x Q matrix, same subjects)}

\item{roi}{region of interest mask}
}
\value{
SPM_T or SPM0D_T object
}
\description{
Paired t-test
}
\examples{
YA <- matrix(rnorm(50), 5, 100)
YB <- YA + matrix(rnorm(50, mean = 0.5), 5, 100)
spm <- ttest_paired(YA, YB)
spmi <- inference(spm, alpha = 0.05, two_tailed = TRUE)
plot_spm(spmi)
}
