% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm1d_r_geom.r
\name{validate_fwhm}
\alias{validate_fwhm}
\title{Validate FWHM estimate}
\usage{
validate_fwhm(fwhm, Q)
}
\arguments{
\item{fwhm}{estimated FWHM}

\item{Q}{number of nodes}
}
\value{
validated FWHM (with warnings if needed)
}
\description{
Validate FWHM estimate
}
